% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchMarking.R
\name{bbm}
\alias{bbm}
\title{Benchmarked Bayes Estimator}
\usage{
bbm(theta_b, w, t, phi, lambda = NA)
}
\arguments{
\item{theta_b}{Bayes estimator.}

\item{w}{Weight vector associates with the benchmark.}

\item{t}{Pre-specified number.}

\item{phi}{Weight vector associates with weighted squared sum of square loss.}

\item{lambda}{Penalty parameter. When it is NULL, lambda = infinity and it will give exact benchmark.}
}
\value{
Resulting Benchmarked Bayes Estimator.
}
\description{
Constrained Bayes estimator which minimizes posterior weighted squared error loss.
}
\references{
\insertRef{datta2011bench}{bayesae}
}
