% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchMarking.R
\name{BM}
\alias{BM}
\title{Benchmarked Bayes Estimator}
\usage{
BM(theta_b, w, t, phi, lambda = NA)
}
\arguments{
\item{theta_b}{Bayes estimator.}

\item{w}{Given weight vector.}

\item{t}{Pre-specified number.}
}
\value{
T square of the input
}
\description{
Constrained Bayes estimator which minimizes posterior weighted squared error loss.
}
\references{
Datta, Gauri Sankar, et al. "Bayesian benchmarking with applications to small area estimation." Test 20.3 (2011): 574-588.
}
