% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchMarking.R
\name{getBench}
\alias{getBench}
\title{Benchmarked bayes estimator}
\usage{
getBench(theta_b, t, interval = NULL, w, phi, lambda = NULL)
}
\arguments{
\item{theta_b}{Bayes estimator.}

\item{t}{Scalar or vector to be benchmarked.}

\item{interval}{95\% credible interval of \eqn{t}. If supplied standard error of \eqn{t} is approximated by \code{range(interval)/4}.}

\item{w}{Weight vector associates with the benchmark.}

\item{phi}{Weight vector associates with weighted squared sum of square loss.}

\item{lambda}{Penalty parameter. Default is \eqn{\lambda = \infty} and it will conduct exact benchmark.}
}
\value{
Resulting Benchmarked Bayes Estimator.
}
\description{
Constrained Bayes estimator which minimizes posterior weighted squared error loss.
}
\references{
\insertRef{datta2011bench}{bayesae}
}
