% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models_fast.R
\name{Model}
\alias{Model}
\title{Stan models}
\usage{
Model(model)
}
\arguments{
\item{model}{Name of a model (\code{"FH"}, \code{"CAR"}, \code{"SAR"}, \code{"MVFH"}).}
}
\description{
Stan models for Fay-Herriot, simultaneous autoregressive and conditional autoregressive models.
}
\references{

}
