% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMVFH.R
\name{BayesMVFH}
\alias{BayesMVFH}
\title{Multivariate Fay-Herriot Model interfacing with Stan}
\arguments{
\item{direct}{\code{m} by \code{s} matrix with direct estimate.}

\item{aux}{Array of auxiliary variable with \code{dim=c(m,s,p)}.}

\item{Di}{\code{m} by \code{s} matrix with sampling variance.}

\item{domain}{\code{m} vector with domain names.}

\item{pars}{Parameters to be monitored.}

\item{iter}{Total iteration.}

\item{warmup}{Warm up. Default is "\code{floor}(\code{iter/2})".}

\item{chains}{Number of chains. Default is 4.}

\item{control}{See the \code{rstan} package document.}

\item{open.progress}{Progress of chiain will be presented if it is \code{TRUE}.}
}
\value{
Simulated posterior sample from the \code{rstan}.
}
\description{
Bayesian approach to \code{s}-variate Fay-Herriot models using \code{stan}.
}
\references{
\insertRef{carpenter2016stan}{bayesae}

\insertRef{guo2016rstan}{bayesae}

\insertRef{vehtari2014waic}{bayesae}
}
