% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMVFH.R
\name{BayesMVFH}
\alias{BayesMVFH}
\title{Multivariate Fay-Herriot Model interfacing with Stan}
\arguments{
\item{direct}{Direct estimates.}

\item{aux}{Auxiliary variables.}

\item{Di}{Pre-specified number.}

\item{domain}{Domain names.}

\item{pars}{Parameters to be monitored.}

\item{iter}{Total iteration.}

\item{warmup}{Warm up. Default is "iter/2".}

\item{chains}{Number of chains. Default is 4.}

\item{control}{See the "rstan" document.}

\item{open.progress}{Progress of chiain will be presented if it is TRUE.}
}
\value{
Simulated posterior sample from the Stan.
}
\description{
Bayesian approach to Multivariate Fay-Herriot models.
}
