% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSAE.R, R/BenchMarking.R
\name{BayesSAE}
\alias{BayesSAE}
\alias{getBench}
\title{Univariate hierarchical Bayes approach to small area estimation.}
\usage{
getBench(theta_b, t, interval = NULL, w, phi, lambda = NULL)
}
\arguments{
\item{theta_b}{Bayes estimator.}

\item{t}{Scalar or vector to be benchmarked.}

\item{interval}{95\% credible interval of \eqn{t}. If supplied standard error of \eqn{t} is approximated by \code{range(interval)/4}.}

\item{w}{Weight vector associates with the benchmark.}

\item{phi}{Weight vector associates with weighted squared sum of square loss.}

\item{lambda}{Penalty parameter. Default is \eqn{\lambda = \infty} and it will conduct exact benchmark.}

\item{formular}{formula}

\item{data}{Data frame with direct estimate and auxiliary variables.}

\item{Di}{\code{m} vector with sampling variance.}

\item{domain}{Vector with Domain names.}

\item{model}{There are three possible models. "FH" for Fay-Herriot model, "CAR" for conditional auto-regressive model and "SAR" for simultaneous auto-regressive model.}

\item{W}{Spatial matrix. If \code{model}="SAR", rowsum should be 1.}

\item{range}{Range of eigenvalues of W (only used if \code{model}="CAR").}

\item{logit.trans}{If true, it transforms simulated theta values to inv.logit(theta).}

\item{pars}{Parameters to be monitored.}

\item{iter}{Total iteration.}

\item{warmup}{Warm up. Default is "\code{floor}(\code{iter}/2)".}

\item{chains}{Number of chains. Default is 4.}

\item{control}{See the \code{rstan} package document.}

\item{open.progress}{Progress of chiain will be presented if it is \code{TRUE}.}
}
\value{
Simulated posterior sample from the \code{rstan}.

Resulting Benchmarked Bayes Estimator.
}
\description{
Hierarchical Bayes Approach to Small Area Estimation interfacing with Stan.

Constrained Bayes estimator which minimizes posterior weighted squared error loss.
}
\references{
\insertRef{carpenter2016stan}{bayesae}

\insertRef{guo2016rstan}{bayesae}

\insertRef{vehtari2014waic}{bayesae}

\insertRef{datta2011bench}{bayesae}
}
