% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSAE.R
\name{BayesSAE}
\alias{BayesSAE}
\title{Univariate hierarchical Bayes approach to small area estimation.}
\arguments{
\item{formula}{formula}

\item{data}{Data frame with direct estimate and auxiliary variables.}

\item{Di}{\code{m} vector with sampling variance.}

\item{domain}{Vector with Domain names.}

\item{model}{There are three possible models. "FH" for Fay-Herriot model, "CAR" for conditional auto-regressive model and "SAR" for simultaneous auto-regressive model.}

\item{W}{Spatial matrix. If \code{model}="SAR", rowsum should be 1.}

\item{logit.trans}{If true, it transforms direct estimate to logit scale and conduct posterior simulation.}

\item{pars}{Parameters to be monitored.}

\item{iter}{Total iteration.}

\item{warmup}{Warm up. Default is "\code{floor}(\code{iter}/2)".}

\item{chains}{Number of chains. Default is 4.}

\item{control}{See the \code{rstan} package document.}

\item{open.progress}{Progress of chiain will be presented if it is \code{TRUE}.}
}
\value{
Simulated posterior sample from the \code{rstan}.
}
\description{
Hierarchical Bayes approach to small area estimation using \code{stan}.
}
\references{
\insertRef{carpenter2016stan}{bayesae}

\insertRef{guo2016rstan}{bayesae}

\insertRef{vehtari2014waic}{bayesae}
}
