% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBench.R
\name{getBench}
\alias{getBench}
\title{Returning penalty parameter for benchmarked Bayes estimator based on supplied sample.}
\usage{
getBench(fit, weight = NULL, par = "theta", interval = NULL)
}
\arguments{
\item{fit}{\code{stan} sample to be benchmarked.}

\item{weight}{Weight vector associates with the benchmark.}

\item{par}{Parameter name to be benchmarked. Default is \eqn{\theta}.}

\item{interval}{95\% credible interval of \eqn{t}. If supplied, other inputs are suppressed and standard error of \eqn{t} is approximated by \code{range(interval)/4}}
}
\description{
Provide penalty parameter for Benchmarked Bayes Estimator.
}
\references{
\insertRef{datta2011bench}{bayesae}
}
